migrated missing functions over
